StatusBarFooter: Implement support for item count
This adds the item count status initialisation and update routines to the
footer component.

Also update call in `MainWindow'. XXX: The value passed in is actually the
filtered list count, which is all items by default but is still not
guaranteed to be the actual total. This is due to `MainWindow' not having
access to the model data.