Remove nonsensical prefixes
The prefixes involving the amount of coin bought, sold, held, or profited
do not make sense, as they are not a function of current price and do not
change on `sync` updates, but only by user action. The user would hence not
need to be notified of their own actions.

* Remove held, bought, sold, made conditions
* WorthChangeCondition: Fix calculation
* ConditionGeneratorTest: Fix test calculations